Actualización de las Excepciones, Update y Delete, Creación de documentatión en swagger,
Actualización de excepciones
Actualización Update y Delete
Dentro de dto, cree una carpeta ownerDTO para almacenar solamente los DTO para el owner
Creación de documentación en Swagger para Owner

+++ src/main/java/org/bedu/veterinaria/service/OwnerService.java
@@ -1,8 +1,10 @@
 package org.bedu.veterinaria.service;
 
-import org.bedu.veterinaria.dto.CreateOwnerDTO;
-import org.bedu.veterinaria.dto.OwnerDTO;
-import org.bedu.veterinaria.dto.UpdateOwnerDTO;
+import org.bedu.veterinaria.dto.ownerDTO.DeleteOwnerDTO;
+import org.bedu.veterinaria.dto.ownerDTO.CreateOwnerDTO;
+import org.bedu.veterinaria.dto.ownerDTO.OwnerDTO;
+import org.bedu.veterinaria.dto.ownerDTO.UpdateOwnerDTO;
+import org.bedu.veterinaria.exception.OwnerNotFoundException;
 import org.bedu.veterinaria.mapper.OwnerMapper;
 import org.bedu.veterinaria.model.Owner;
 import org.bedu.veterinaria.repository.OwnerRepository;
@@ -10,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class OwnerService {
@@ -24,28 +27,54 @@ public class OwnerService {
         this.mapper = mapper;
     }
 
+    //Obtener todos los clientes
     public List<OwnerDTO> findAll(){
         return repository.findAll().stream().map(mapper::toDTO).toList();
     }
 
+    //Crear un nuevo cliente
     public OwnerDTO save(CreateOwnerDTO data){
         Owner entity = repository.save(mapper.toModel(data));
         return mapper.toDTO(entity);
     }
 
-    public OwnerDTO updateById(Long idOwner, UpdateOwnerDTO data){
-        data.setIdOwner(idOwner);
-        Owner entity = repository.save(mapper.toModel(data));
-        return mapper.toDTO(entity);
+    //Actualiza un cliente existente
+    public void updateOwner(long idOwner, UpdateOwnerDTO data) throws OwnerNotFoundException{
+        Optional<Owner> result = repository.findById(idOwner);
+        if(!result.isPresent()){
+            throw new OwnerNotFoundException(idOwner);
+        }
+        Owner owner = result.get();
+        mapper.updateOwner(owner,data);
+        repository.save(owner);
     }
 
-    public Boolean deleteOwner(Long idOwner){
-        try {
-            repository.deleteById(idOwner);
-            return true;
-        } catch (Exception e){
-            return false;
+//    public OwnerDTO updateById(Long idOwner, UpdateOwnerDTO data){
+//        data.setIdOwner(idOwner);
+//        Owner entity = repository.save(mapper.toModel(data));
+//        return mapper.toDTO(entity);
+//    }
+
+    //Borra un cliente
+
+    public void deleteOwner(long idOwner, DeleteOwnerDTO data) throws OwnerNotFoundException{
+        Optional<Owner> result = repository.findById(idOwner);
+        if(!result.isPresent()){
+            throw new OwnerNotFoundException(idOwner);
         }
+        Owner entity = result.get();
+        mapper.deleteOwner(entity,data);
+        repository.deleteById(idOwner);
     }
+
+
+//    public Boolean deleteOwner(Long idOwner){
+//        try {
+//            repository.deleteById(idOwner);
+//            return true;
+//        } catch (Exception e){
+//            return false;
+//        }
+//    }
 
 }
